Many changes
Many changes getting a test build ready, added loading screen and fixed any bugs i could see.

-Loading screen required a async loading system to be implemented.

-Also many fixes to UI and any bugs

+++ Assets/Scripts/Objects/SavingObj/Trees/TreeObject.cs
@@ -60,7 +60,7 @@ public class TreeObject : SaveableObject
     public override void Load(string[] values)
     {
         //Getting data from saved object in list, first variable is index 4 of values array, if another varaible is added it would be index 5 and so on
-        timeToGrow = int.Parse(values[4]);
+        timeToGrow = float.Parse(values[4]);
         base.Load(values);
     }
 }
